Rename pkg/util/netsh to pkg/util/netifcmd now that it supports more than netsh
