+++ pkg/proxy/winuserspace/BUILD
@@ -22,7 +22,7 @@ go_library(
         "//pkg/proxy/config:go_default_library",
         "//pkg/proxy/util:go_default_library",
         "//pkg/util/ipconfig:go_default_library",
-        "//pkg/util/netsh:go_default_library",
+        "//pkg/util/netifcmd:go_default_library",
         "//pkg/util/slice:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/types:go_default_library",
@@ -43,7 +43,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//pkg/proxy:go_default_library",
-        "//pkg/util/netsh/testing:go_default_library",
+        "//pkg/util/netifcmd/testing:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/api/discovery/v1beta1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
