+++ pkg/util/netifcmd/doc.go
@@ -14,5 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package netsh provides an interface and implementations for running Windows netsh commands.
-package netsh // import "k8s.io/kubernetes/pkg/util/netsh"
+// Package netifcmd provides an interface and implementations for Windows and Darwin.
+package netifcmd // import "k8s.io/kubernetes/pkg/util/netifcmd"
