+++ pkg/util/netifcmd/netsh.go
@@ -1,4 +1,4 @@
-package netsh
+package netifcmd
 
 import (
 	"fmt"
