+++ pkg/util/netifcmd/testing/BUILD
@@ -8,8 +8,8 @@ load(
 go_library(
     name = "go_default_library",
     srcs = ["fake.go"],
-    importpath = "k8s.io/kubernetes/pkg/util/netsh/testing",
-    deps = ["//pkg/util/netsh:go_default_library"],
+    importpath = "k8s.io/kubernetes/pkg/util/netifcmd/testing",
+    deps = ["//pkg/util/netifcmd:go_default_library"],
 )
 
 filegroup(
